apps/server/pyproject.toml
[project]
name = "nush-server"
version = "0.1.0"
description = "nush Flask API"
requires-python = ">=3.11"
dependencies = [
  "flask>=3.0.0",
  "flask-sqlalchemy>=3.1.1",
  "flask-migrate>=4.0.5",
  "flask-cors>=4.0.0",
  "flask-limiter>=3.7.0",
  "bcrypt>=4.0.1",
  "faker>=24.0.0",
  "python-dotenv>=1.0.1",
  "gunicorn>=21.2.0",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "stripe>=10.0.0"
]

[project.optional-dependencies]
dev = ["ruff>=0.6.2"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
